+++ src/Clang/Version/Internal/Check.hs
@@ -1,0 +1,8 @@
+{-# LANGUAGE TemplateHaskell #-}
+
+-- | Use 'checkUserClangVersion' to process the @CLANG_VERSION@ macro
+module Clang.Version.Internal.Check () where
+
+import Clang.Version.Internal (checkUserClangVersion)
+
+$(checkUserClangVersion)